web: add gunicorn app

+++ WEB/init.sh
@@ -1,4 +1,7 @@
-#!/bin/sh
+#!/bin/bash
+
+cd "$( realpath "$(dirname "$0")" )" || exit 1
+FROMHERE="$( pwd )"
 
 rm -rf /home/box/web/*
 mkdir -p /home/box/web/public
@@ -8,7 +11,13 @@ mkdir -p /home/box/web/public/js
 mkdir -p /home/box/web/uploads
 mkdir -p /home/box/web/etc
 
-BASEDIR=$(dirname "$0")
-unlink /etc/nginx/sites-enabled/default
-ln -fs "$(realpath "$BASEDIR")/nginx.conf"  /etc/nginx/sites-enabled/test.conf
-/etc/init.d/nginx restart
+if [ -f /etc/nginx/sites-enabled/default ]; then
+    unlink /etc/nginx/sites-enabled/default
+fi
+
+ln -fs "$FROMHERE/nginx.conf" /etc/nginx/sites-enabled/test.conf
+systemctl restart nginx.service
+
+cp "hello.py" /home/box/web/hello.py
+
+gunicorn -b 0.0.0.0:8080 -w 2 "hello:application"
